+++ _define.php
@@ -25,7 +25,7 @@ $this->registerModule(
 	/* Name */		    "dcCKEditorAddons",
 	/* Description*/	"Add CKEditor plugins easily to your blog",
 	/* Author */		"Nicolas Roudaire",
-	/* Version */		'0.2.1',
+	/* Version */		'0.3.0',
     array('permissions' => 'contentadmin,pages',
           'type' => 'plugin',
           'dc_min' => '2.7',