+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+dcCKEditorAddons 0.3.0 - 2015-08-17
+===================================
+* Manage actions with drop/down list
+* Allow addons to be deleted.
+
+dcCKEditorAddons 0.2.0 - 2015-08-14
+===================================
+* Display dependencies
+* Retrieve buttons names
+
 dcCKEditorAddons 0.1.0 - 2015-08-13
 ===================================
 * First public release